plus de warning pour les Matrix

+++ src/Fraction.h
@@ -13,18 +13,18 @@ class Fraction
 
 private:
 
-    int numerateur, denominateur;
+    long int numerateur, denominateur;
 
 public:
 
     Fraction();
-    Fraction(const int & n);
-    Fraction(const int & n, const int & m);
+    Fraction(const long int & n);
+    Fraction(const long int & n, const long int & m);
     void afficher(std::ostream& out) const;
     bool estEgal(const Fraction & f) const;
     bool estPlusPetitQue(const Fraction & f) const;
     Fraction& operator+=(const Fraction & f);
-    int pgcd(int a, int b);
+    long int pgcd(long int a, long int b);
     void simplifie();
     Fraction double2Fraction(const double & f);
